starting KnightRider class. still odd nil in Characters
